Integer cleanup

+++ include/base32.h
@@ -1,11 +1,13 @@
+#include <stddef.h>
+
 #ifndef _b32_h_
 #define _b32_h_
 
-int base32_encode_length(int rawLength);
-int base32_decode_length(int base32Length);
-void base32_encode_into(const void *_buffer, unsigned int bufLen, char *base32Buffer);
-char *base32_encode(const void *buf, unsigned int len);
-int base32_decode_into(const char *base32Buffer, unsigned int base32BufLen, void *_buffer);
-void *base32_decode(const char *buf, unsigned int *outlen);
+size_t base32_encode_length(size_t rawLength);
+size_t base32_decode_length(size_t base32Length);
+void base32_encode_into(const void *_buffer, size_t bufLen, char *base32Buffer);
+char *base32_encode(const void *buf, size_t len);
+size_t base32_decode_into(const char *base32Buffer, size_t base32BufLen, void *_buffer);
+void *base32_decode(const char *buf, size_t *outlen);
 #endif
